[chore]: 패키징 잘못된 부분 수정 (#12)
