+++ app/src/main/java/com/konkuk/codion/data/dto/request/LoginRequest.kt
@@ -1,0 +1,9 @@
+package com.konkuk.codion.data.dto.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LoginRequest(
+    val email: String,
+    val password: String
+)
